fix: scheduler thread should be a daemon (#125)
Because of the typo scheduler thread was not treated as a daemon thread
- therefore it wouldn't exit once the main thread exitted, but would
wait for itself to finish. The process needs to exit fully to reload the
configuration (for example, an input is disabled or log level changed).
This was usually not a problem, as scheduler was able to exit by itself
in most cases. But in some cases, when there were more tasks to schedule
than task_queue_size, the scheduler would be stuck in a loop and never
exit.
This change will make it so that scheduler thread always gets killed
once the main thread exits.